requirements.txt
streamlit>=1.37.0
pandas>=2.2.0
numpy>=1.26.0
google-generativeai>=0.7.0
rapidfuzz>=3.9.0
python-dotenv>=1.0.0

